error with mongodb

+++ ReceiveRawData.java
@@ -29,8 +29,8 @@ public class ReceiveRawData{
 			System.out.println("Date is " +date+" at "+ time);
 
 			CPU.getInstance().addTask(new Task( data[0], 
-																					Double.parseDouble(data[1]),
-																					Double.parseDouble(data[2]),
+																					Float.parseFloat(data[1]),
+																					Float.parseFloat(data[2]),
 																					Float.parseFloat(data[3]),
 																					date,
 																					frameTemp));
@@ -39,11 +39,11 @@ public class ReceiveRawData{
 			db = mongoClient.getDB("hcm_traffic");
 			DBCollection gpsData = db.getCollection("GPSdata");
 			BasicDBObject doc = new BasicDBObject("device_id", data[0]).
-																append("latitude", Double.parseDouble(data[1])).
-																append("longitude", Double.parseDouble(data[2])).
+																append("latitude", Float.parseFloat(data[1])).
+																append("longitude", Float.parseFloat(data[2])).
 																append("speed", Float.parseFloat(data[3])).
-																append("reliability", Double.parseDouble(data[4])).
-																append("number_of_satellites", Double.parseDouble(data[5])).
+																append("reliability", Float.parseFloat(data[4])).
+																append("number_of_satellites", Float.parseFloat(data[5])).
 																append("current_time", new BasicDBObject("date", date).append("frame", frameTemp).append("time", time));
 	    gpsData.insert(doc);
 		} catch (UnknownHostException e) {
